+++ src/TestableHttpClient/TestableHttpMessageHandler.cs
@@ -17,9 +17,18 @@ public class TestableHttpMessageHandler : HttpMessageHandler
     /// </summary>
     public IEnumerable<HttpRequestMessage> Requests => httpRequestMessages;
 
+    protected override void Dispose(bool disposing)
+    {
+        DisposeRequestMessages();
+        base.Dispose(disposing);
+    }
+
+    [System.Diagnostics.CodeAnalysis.SuppressMessage("Reliability", "CA2000:Dispose objects before losing scope", Justification = "It gets disposed in the dispose method")]
     protected override async Task<HttpResponseMessage> SendAsync(HttpRequestMessage request, CancellationToken cancellationToken)
     {
-        httpRequestMessages.Enqueue(request);
+        Guard.ThrowIfNull(request);
+
+        httpRequestMessages.Enqueue(await HttpRequestMessageCloner.ClonaAsync(request, cancellationToken).ConfigureAwait(false));
 
         HttpResponseMessage responseMessage = new();
         HttpResponseContext context = new(request, httpRequestMessages, responseMessage, Options);
@@ -27,9 +36,9 @@ public class TestableHttpMessageHandler : HttpMessageHandler
 
         responseMessage.RequestMessage ??= request;
 
-#if !NET6_0_OR_GREATER
+        // In .NET Standard, a response message can be null, but we need it to be at least empty like in the newer versions.
+        // Newer versions will always have a content, even if it's empty.
         responseMessage.Content ??= new StringContent("");
-#endif
 
         return responseMessage;
     }
@@ -55,6 +64,15 @@ public class TestableHttpMessageHandler : HttpMessageHandler
     /// <remarks>The configuration itself (Options and the configured IResponse) will not be cleared or reset.</remarks>
     public void ClearRequests()
     {
+        DisposeRequestMessages();
         httpRequestMessages.Clear();
+    }
+
+    private void DisposeRequestMessages()
+    {
+        foreach (HttpRequestMessage request in httpRequestMessages)
+        {
+            request.Dispose();
+        }
     }
 }