+++ test/TestableHttpClient.IntegrationTests/AssertingRequests.cs
@@ -162,17 +162,31 @@ public class AssertingRequests
         using StringContent content = new("my special content");
         _ = await client.PostAsync("https://httpbin.org/post", content, TestContext.Current.CancellationToken);
 
-#if NETFRAMEWORK
-        // On .NET Framework the HttpClient disposes the content automatically. So we can't perform the same test.
-        testHandler.ShouldHaveMadeRequests();
-#else
         testHandler.ShouldHaveMadeRequests().WithContent("my special content");
         testHandler.ShouldHaveMadeRequests().WithContent("my*content");
         testHandler.ShouldHaveMadeRequests().WithContent("*");
 
         Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithContent(""));
         Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithContent("my"));
-#endif
+    }
+
+    [Fact]
+    public async Task AssertingContent_WhenOriginalContentIsDisposed()
+    {
+        using TestableHttpMessageHandler testHandler = new();
+        using HttpClient client = new(testHandler);
+
+        using (StringContent content = new("my special content"))
+        {
+            _ = await client.PostAsync("https://httpbin.org/post", content, TestContext.Current.CancellationToken);
+        }
+
+        testHandler.ShouldHaveMadeRequests().WithContent("my special content");
+        testHandler.ShouldHaveMadeRequests().WithContent("my*content");
+        testHandler.ShouldHaveMadeRequests().WithContent("*");
+
+        Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithContent(""));
+        Assert.Throws<HttpRequestMessageAssertionException>(() => testHandler.ShouldHaveMadeRequests().WithContent("my"));
     }
 
     [Fact]
@@ -184,12 +198,7 @@ public class AssertingRequests
         using StringContent content = new("{}", Encoding.UTF8, "application/json");
         _ = await client.PostAsync("https://httpbin.org/post", content, TestContext.Current.CancellationToken);
 
-#if NETFRAMEWORK
-        // On .NET Framework the HttpClient disposes the content automatically. So we can't perform the same test.
-        testHandler.ShouldHaveMadeRequests();
-#else
         testHandler.ShouldHaveMadeRequests().WithJsonContent(new { });
-#endif
     }
 
     [Fact]